Added: Ability to edit the existing contact by their name

+++ AddressBookSystem/Contacts.java
@@ -73,7 +73,7 @@ public class Contacts {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(long phoneNumber) {
+    public void setPhoneNumber(int phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
